+++ question/type/drawing/drawingarea.php
@@ -38,12 +38,13 @@ $stid = optional_param('stid', 0, PARAM_INT);
 $attemptid = required_param('attemptid', PARAM_RAW_TRIMMED);
 $uniquefieldnameattemptid = required_param('uniquefieldnameattemptid', PARAM_RAW_TRIMMED);
 $sesskey = required_param('sesskey', PARAM_RAW);
+$attemptcount = optional_param('attemptcount', 1, PARAM_INT);
 
-if (!confirm_sesskey()) {
-    die();
+if (!confirm_sesskey($sesskey)) {
+    die("Session lost");
 }
 if (!$fhd = $DB->get_record('qtype_drawing', array('questionid' => $id))) {
-    print_error("No such question.");
+    die("No such question.");
 }
 $reducedmode = 0;
 $displaystyle = '';
@@ -102,10 +103,11 @@ if (has_capability('mod/quiz:grade', context::instance_by_id($question->contexti
          var qtype_drawing_str_saving = "<?php print_string("saving", "qtype_drawing");?>";
          var qtype_drawing_str_saveannotation = "<?php print_string("saveannotation", "qtype_drawing");?>";
          var questionid = <?php echo $id;?>;
-         var sesskey = '<?php echo $sesskey;?>';
+         var sesskey = '<?php echo strip_tags($sesskey);?>';
          var stid = <?php echo $stid;?>;
-         var attemptid = '<?php echo $attemptid;?>';
-         var uniquefieldnameattemptid = '<?php echo $uniquefieldnameattemptid;?>';
+         var attemptid = '<?php echo strip_tags($attemptid);?>';
+         var attemptcount = '<?php echo $attemptcount;?>';
+         var uniquefieldnameattemptid = '<?php echo strip_tags($uniquefieldnameattemptid);?>';
       </script>
       <script type="text/javascript"
               src="<?php echo $CFG->wwwroot.'/question/type/drawing/';?>lib/jquery.js"></script>
@@ -476,7 +478,7 @@ if (has_capability('mod/quiz:grade', context::instance_by_id($question->contexti
                               text-align:left">
                         <?php
                         echo '<ul id="listofannotaions">';
-                        $fields = array('questionid' => $question->id, 'attemptid' => $attemptid, 'annotatedfor' => $stid);
+                        $fields = array('questionid' => $question->id, 'attemptid' => $attemptid, 'annotatedfor' => $stid, 'attemptcount' => $attemptcount);
                         if ($annotations = $DB->get_records('qtype_drawing_annotations', $fields, 'timemodified DESC')) {
                             foreach ($annotations as $teacherannotation) {
                                 $user = $DB->get_record('user', array('id' => $teacherannotation->annotatedby));