+++ question/type/drawing/saveannotation.php
@@ -38,6 +38,7 @@ $sesskey = required_param('sesskey', PARAM_RAW);
 $stid = required_param('stid', PARAM_INT);
 $attemptid = required_param('attemptid', PARAM_RAW_TRIMMED);
 $annotation = required_param('annotation', PARAM_RAW);
+$attemptcount = optional_param('attemptcount', 1, PARAM_INT);
 
 if (!confirm_sesskey()) {
     echo json_encode(array('result' => 'Session lost.'));
@@ -58,8 +59,15 @@ if (!$fhd = $DB->get_record('qtype_drawing', array('questionid' => $id))) {
     die();
 }
 
+// Just in case, remove any <script> if direct saving happens (how?!).
+$annotation = preg_replace('#<script(.*?)>(.*?)</script>#is', '', $annotation);
+// After cleaning from script, is it empty?.
+if(trim($annotation) == ''){
+    echo json_encode(array('result' => 'No annotation submitted'));
+    die();
+}
 // Check if record exists.
-$fields = array('questionid' => $id, 'annotatedby' => $USER->id, 'annotatedfor' => $stid, 'attemptid' => $attemptid);
+$fields = array('questionid' => $id, 'annotatedby' => $USER->id, 'annotatedfor' => $stid, 'attemptid' => $attemptid, 'attemptcount' => $attemptcount);
 if ($recordexists = $DB->get_record('qtype_drawing_annotations', $fields)) {
     // Update annotation.
     $annotationrecord = new stdClass();
@@ -70,6 +78,7 @@ if ($recordexists = $DB->get_record('qtype_drawing_annotations', $fields)) {
     $annotationrecord->annotatedby = $USER->id;
     $annotationrecord->annotatedfor = $stid;
     $annotationrecord->attemptid = $attemptid;
+    $annotationrecord->attemptcount = $attemptcount;
     $annotationrecord->notes = '';
     $DB->update_record('qtype_drawing_annotations', $annotationrecord);
 } else {
@@ -83,8 +92,10 @@ if ($recordexists = $DB->get_record('qtype_drawing_annotations', $fields)) {
     $annotationrecord->annotatedby = $USER->id;
     $annotationrecord->annotatedfor = $stid;
     $annotationrecord->attemptid = $attemptid;
+    $annotationrecord->attemptcount = $attemptcount;
     $annotationrecord->notes = '';
     $DB->insert_record('qtype_drawing_annotations', $annotationrecord);
+    $result = 'insert '.$attemptcount;
 }
 $result = 'OK';
 echo json_encode($result);